# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aura Render API",
    "version": "1.0.0",
    "description": "\n# Aura Render API Documentation\n\nAura Render is an AI-powered video generation platform that transforms text into professional videos.\n\n## Features\n- **AI Content Analysis**: Intelligent analysis of input text for optimal video generation\n- **Material Matching**: Automatic matching and acquisition of relevant video materials\n- **Real-time Processing**: WebSocket-based real-time progress tracking\n- **Multi-format Support**: Support for various video formats and quality levels\n- **Enterprise Security**: JWT authentication, rate limiting, and audit logging\n- **Batch Processing**: Handle multiple video generation requests efficiently\n\n## Authentication\nMost endpoints require authentication using JWT tokens. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## Rate Limiting\nAPI requests are subject to rate limiting:\n- **Free Users**: 10 requests/minute, 600 requests/hour\n- **Premium Users**: 100 requests/minute, 6000 requests/hour\n- **Enterprise**: Custom limits available\n\n## Error Handling\nAll errors follow a consistent format:\n```json\n{\n  \"error\": {\n    \"code\": \"ERROR_CODE\",\n    \"message\": \"Human-readable error message\",\n    \"details\": {\n      \"field\": \"Additional error context\"\n    }\n  }\n}\n```\n        ",
    "termsOfService": "https://aurarender.com/terms",
...
    },
...
    }
...
  "servers": [
...
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login endpoint"
      }
    },
    "schemas": {
      "UserLogin": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "securePassword123"
          },
          "remember_me": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refresh_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "example": 3600
          }
        }
      },
...
      "VideoGenerationRequest": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10000,
            "example": "Create an engaging promotional video about our new AI-powered productivity tool that helps teams collaborate more effectively.",
            "description": "The text content to convert into video"
          },
          "video_type": {
            "type": "string",
            "enum": [
              "promotional",
              "educational",
              "social",
              "presentation",
              "story"
            ],
            "example": "promotional",
            "description": "Type of video to generate"
          },
          "style": {
            "type": "object",
            "properties": {
              "theme": {
                "type": "string",
                "enum": [
                  "corporate",
                  "modern",
                  "playful",
                  "elegant",
                  "tech"
                ],
                "example": "modern"
              },
              "color_scheme": {
                "type": "string",
                "example": "blue-gradient"
              },
              "font_family": {
                "type": "string",
                "example": "Roboto"
              }
            }
          },
          "audio": {
            "type": "object",
            "properties": {
              "voice": {
                "type": "string",
                "enum": [
                  "xiaoyun",
                  "xiaogang",
                  "xiaoxiao",
                  "xiaomeng"
                ],
                "example": "xiaoyun",
                "description": "Text-to-speech voice selection"
              },
              "speed": {
                "type": "number",
                "minimum": 0.5,
                "maximum": 2.0,
                "example": 1.0,
                "description": "Speech speed multiplier"
              },
              "background_music": {
                "type": "boolean",
                "example": true,
                "description": "Whether to add background music"
              }
            }
          },
          "video_config": {
            "type": "object",
            "properties": {
              "resolution": {
                "type": "string",
                "enum": [
                  "1920x1080",
                  "1280x720",
                  "3840x2160"
                ],
                "example": "1920x1080"
              },
              "fps": {
                "type": "integer",
                "enum": [
                  24,
                  30,
                  60
                ],
                "example": 30
              },
              "duration_limit": {
                "type": "integer",
                "minimum": 10,
                "maximum": 300,
                "example": 60,
                "description": "Maximum video duration in seconds"
              },
              "format": {
                "type": "string",
                "enum": [
                  "mp4",
                  "mov",
                  "avi"
                ],
                "example": "mp4"
              }
            }
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "normal",
              "high"
            ],
            "example": "normal",
            "description": "Processing priority (premium feature)"
          }
        }
      },
      "VideoGenerationResponse": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "example": "task_abcd1234efgh5678",
            "description": "Unique task identifier for tracking"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "completed",
              "failed"
            ],
            "example": "queued"
          },
          "estimated_completion": {
            "type": "string",
            "format": "date-time",
            "example": "2023-12-01T10:05:00Z"
          },
          "websocket_url": {
            "type": "string",
            "example": "wss://api.aurarender.com/ws/tasks/task_abcd1234efgh5678",
            "description": "WebSocket URL for real-time progress updates"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid request parameters"
              },
              "details": {
                "type": "object",
                "example": {
                  "text": "Text content is required"
                }
              }
            }
          }
        }
      },
      "RateLimitError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "RATE_LIMIT_EXCEEDED"
              },
              "message": {
                "type": "string",
                "example": "Rate limit exceeded"
              },
              "details": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "integer",
                    "example": 100
                  },
                  "remaining": {
                    "type": "integer",
                    "example": 0
                  },
                  "reset_time": {
                    "type": "integer",
                    "example": 1701432000
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "User login",
        "description": "Authenticate user and get JWT tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too many login attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh JWT token",
        "description": "Get a new access token using refresh token",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refresh_token"
                ],
                "properties": {
                  "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "User logout",
        "description": "Invalidate current session tokens",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logged out successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/video/generate": {
      "post": {
        "tags": [
          "Video Generation"
        ],
        "summary": "Generate video from text",
        "description": "Create a video from text content using AI processing",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoGenerationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Video generation task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoGenerationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}